remove parser class
